+++ src/completions/loopObjCompletion.ts
@@ -1,12 +1,7 @@
 import * as vscode from 'vscode'
-import completionItem from '../modules/completionItem'
-import loopInfo from '../static/info/loopInfo'
-import Cursor from '../modules/Cursor'
-
-const IGNORE_REG = /{{--\s*@(each|for)/g
-const START_REG = /@(each|for)/g
-const END_REG = /@end/g
-const LOOP_VAR_REG = /\bloop\./g
+import { completionItem } from '../modules/completionItem'
+import { loopInfo as info } from '../static/info/loopInfo'
+import { Cursor } from '../modules/Cursor'
 
 const triggerChars = ['.', 'l']
 
@@ -19,30 +14,29 @@ export default vscode.languages.registerCompletionItemProvider(
         ): vscode.CompletionItem[] {
             const cursor = new Cursor(pos, doc)
 
-            if (cursor.notBetween(START_REG, END_REG, IGNORE_REG)) {
+            if (!cursor.isInsideLoop()) {
                 return []
             }
 
             if (cursor.prevCharIs('l')) {
                 const field = vscode.CompletionItemKind.Variable
-                return [completionItem('loop', loopInfo.loop, field)]
+                return [completionItem('loop', info.loop, field)]
             }
 
             const range = new vscode.Range(pos.with(pos.line, 0), pos)
-            const textBefore = doc.getText(range)
-            const match = LOOP_VAR_REG.test(textBefore.trim())
+            const textBefore = doc.getText(range).trimEnd()
 
-            if (!match) {
+            if (!textBefore.endsWith('loop.')) {
                 return []
             }
 
             const field = vscode.CompletionItemKind.Field
 
             return [
-                completionItem('index', loopInfo.index, field),
-                completionItem('first', loopInfo.first, field),
-                completionItem('last', loopInfo.last, field),
-                completionItem('iter', loopInfo.iter, field),
+                completionItem('index', info.index, field),
+                completionItem('first', info.first, field),
+                completionItem('last', info.last, field),
+                completionItem('iter', info.iter, field),
             ]
         },
     },