+++ src/modules/Cursor.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode'
 
-export default class Cursor {
+export class Cursor {
     public constructor(private pos: vscode.Position, private doc: vscode.TextDocument) {}
 
     public prevChar(): string {
@@ -29,8 +29,28 @@ export default class Cursor {
         return lastStart && (!lastEnd || lastStart.index > lastEnd.index)
     }
 
-    public notBetween(startReg: RegExp, endReg: RegExp, ignoreReg: RegExp): boolean {
-        return !this.isBetween(startReg, endReg, ignoreReg)
+    public isInsideBraces(): boolean {
+        const IGNORE = /\\\{\{/g
+        const START = /\{\{/g
+        const END = /\}\}/g
+
+        return this.isBetween(START, END, IGNORE)
+    }
+
+    public isInsideLoop(): boolean {
+        const IGNORE = /{{--\s*@(each|for)/g
+        const START = /@(each|for)/g
+        const END = /@end/g
+
+        return this.isBetween(START, END, IGNORE)
+    }
+
+    public isInsideComponent(): boolean {
+        const IGNORE = /{{--\s*@component/g
+        const START = /@component/g
+        const END = /@end/g
+
+        return this.isBetween(START, END, IGNORE)
     }
 
     private textBefore(): string {