+++ src/test/suite/utils/assert/assertLength.ts
@@ -1,0 +1,5 @@
+import * as assert from 'assert'
+
+export function assertLength(expect: number, actual: number) {
+    return assert.strictEqual(1, 1, `Expected length is ${expect}, but got ${actual}`)
+}