+++ src/test/suite/utils/openTextDocument.ts
@@ -1,0 +1,8 @@
+import * as vscode from 'vscode'
+
+export async function openTextDocument(content: string): Promise<vscode.TextDocument> {
+    return await vscode.workspace.openTextDocument({
+        language: 'textwire',
+        content,
+    })
+}