+++ src/test/suite/utils/triggerCompletion.ts
@@ -1,0 +1,18 @@
+import * as vscode from 'vscode'
+
+export async function triggerCompletion(
+    pos: vscode.Position,
+    uri: vscode.Uri,
+): Promise<vscode.CompletionItem[]> {
+    const completions = await vscode.commands.executeCommand<vscode.CompletionList>(
+        'vscode.executeCompletionItemProvider',
+        uri,
+        pos,
+    )
+
+    if (!completions) {
+        return []
+    }
+
+    return completions.items
+}