+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.5.3] - 2025-02-02
+- Added tests coverage to test all the necessary features
+- Change so that you can only see `@break`, `@breakIf`, `@continue`, `@continueIf` directives when you are inside a loop
+- Change so that you can only see `@slot` and `@slot(name)` directives when you are inside a component
+
 ## [1.5.2] - 2025-01-30
 - Fixed issue where loop object would only show autocomplete suggestions when it was inside `{{` and `}}` brackets
 - Refactor code to make it more readable and maintainable